+++ projects/ng-form-foundry/src/public-api.ts
@@ -11,5 +11,6 @@ export * from './lib/dynamic-recursive-form/leaf-list-renderer/leaf-list-rendere
 export * from './lib/dynamic-recursive-form/leaf-enum-renderer/leaf-enum-renderer.component';
 export * from './lib/dynamic-recursive-form/radix-input/radix-input.directive';
 export * from './lib/directives/overflow-tooltip.directive';
+export * from './lib/directives/rich-tooltip.directive';
 export * from './lib/dynamic-recursive-form/node-group-list-renderer/node-group-list-renderer.component';
 export * from './lib/dynamic-recursive-form/node-map-renderer/node-map-renderer.component';